\name{genAR}
\alias{genAR}
\title{Generates accessores and replacements for classes}
\arguments{
  \item{object}{a class object}
}
\value{
  generates the accessors and replacement methods
}
\description{
  Generates accessores and replacements for classes
}
\author{
  EJ \email{ernesto.jardim@jrc.ec.europa.eu}
}

