\name{rngyear<-}
\alias{rngyear<-}
\title{year range replacement}
\arguments{
  \item{object}{a \code{a4aM} object}

  \item{value}{a \code{vector} with max and min year range}
}
\value{
  a \code{a4aM} object
}
\description{
  method to replace \code{a4aM} object's year range.
}
\author{
  EJ \email{ernesto.jardim@jrc.ec.europa.eu}
}

