\name{valida4aM}
\alias{valida4aM}
\title{====================================================================
M
====================================================================}
\usage{
  valida4aM(object)
}
\description{
  ====================================================================
  M
  ====================================================================
}

