\name{vcov<-}
\alias{vcov<-}
\title{set (co)variance}
\arguments{
  \item{object}{a \code{a4aGr} object}

  \item{value}{a \code{array} object}
}
\value{
  a \code{a4aGr} object
}
\description{
  set method for \code{a4aGr} object's slot \code{vcov}.
}
\author{
  EJ \email{ernesto.jardim@jrc.ec.europa.eu}
}

